# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daton Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization details
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
  /organizations/{orgId}/users:
    get:
      operationId: listOrgUsers
      tags: [organizations]
      summary: List users in organization with roles and module permissions
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgUser"
                required:
                  - users
    post:
      operationId: createOrgUser
      tags: [organizations]
      summary: Create a user directly in the organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrgUserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /organizations/{orgId}/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [organizations]
      summary: Update a user's role
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [operator, analyst]
              required:
                - role
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /organizations/{orgId}/users/{userId}/modules:
    put:
      operationId: updateUserModules
      tags: [organizations]
      summary: Set a user's module permissions
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                modules:
                  type: array
                  items:
                    type: string
              required:
                - modules
      responses:
        "200":
          description: Modules updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  modules:
                    type: array
                    items:
                      type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nomeFantasia:
          type: ["string", "null"]
        cnpj:
          type: ["string", "null"]
        inscricaoEstadual:
          type: ["string", "null"]
        dataFundacao:
          type: ["string", "null"]
        statusOperacional:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        nomeFantasia:
          type: ["string", "null"]
        cnpj:
          type: ["string", "null"]
        inscricaoEstadual:
          type: ["string", "null"]
        dataFundacao:
          type: ["string", "null"]
        statusOperacional:
          type: ["string", "null"]
...
    OrgUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        modules:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - modules
...
    CreateOrgUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [org_admin, operator, analyst]
        modules:
          type: array
          items:
            type: string
            enum: [documents, legislations, employees, units, departments, positions]
      required:
        - name
        - email
        - password
        - role
        - modules
    CreateOrgUserResponse:
      allOf:
        - $ref: "#/components/schemas/OrgUser"
...