# lib/api-zod/src/generated/types/updateOrganizationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daton Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrganizationBody {
  name?: string;
  /** @nullable */
  nomeFantasia?: string | null;
  /** @nullable */
  cnpj?: string | null;
  /** @nullable */
  inscricaoEstadual?: string | null;
  /** @nullable */
  dataFundacao?: string | null;
  /** @nullable */
  statusOperacional?: string | null;
}